+++ src/components/NavBar/NavBar.module.scss
@@ -4,7 +4,7 @@ body {
 
 .nav {
   background-color: #526578;
-  color: #FFFFFF;
+  color: #ffffff;
   display: flex;
   height: 64px;
   gap: 2rem;
@@ -40,7 +40,7 @@ body {
 }
 
 .nav li.active {
-  background-color: #F4F4F4;
+  background-color: #f4f4f4;
   color: #000000;
 }
 
@@ -69,7 +69,7 @@ body {
   border: none;
   margin-left: 1rem;
   color: #000000;
-  background: #9AA8B5;
+  background: #9aa8b5;
   border-radius: 12px;
 }
