--- src/styles/navbar/navbar.sass
@@ -1,78 +1,0 @@
-@import "./variables"
-
-body
-    margin: 0
-
-.nav
-    background-color: $primaryColor
-    color: $navTextColor
-    display: flex
-    height: $navbarHeight
-    gap: 2rem
-    padding: 0 1rem
-    justify-content: center
-
-.nav ul
-    display: flex
-    padding: 0 1rem
-    list-style: none
-    justify-content: space-between
-    width: 100%
-
-.nav p
-    align-items: center
-    justify-content: center
-    padding: 0
-    margin: 0
-    list-style: none
-    display: flex
-
-.nav a
-    color: inherit
-    width: $navLinkWidth
-    height: 100%
-    justify-content: center
-    text-decoration: none
-    display: flex
-    align-items: center
-
-.nav li.active
-    background-color: $activeColor
-    color: $activeTextColor
-
-.user
-    display: flex
-    height: $navbarHeight
-    align-items: center
-    gap: 1rem
-    overflow: hidden
-    white-space: nowrap
-
-.leftPart
-    display: flex
-    height: $navbarHeight
-    overflow: hidden
-    white-space: nowrap
-
-.logout
-    display: flex
-    justify-content: center
-    align-items: center
-    width: $logoutBtnWidth
-    height: $logoutBtnHeight
-    border: none
-    margin-left: 1rem
-    color: $logoutBtnTextColor
-    background: $logoutBtnColor
-    border-radius: 12px
-
-.nav li:hover:not(.active)
-    background-color: $hoverColor
-
-.nav span
-    padding-left: 0.5rem
-
-.nav .iconLink
-    display: flex
-    height: $navbarHeight
-    align-items: center
