--- src/styles/navbar/variables.sass
@@ -1,22 +1,0 @@
-// Navbar
-$primaryColor: #526578
-$navbarHeight: 64px
-
-// Nav link
-$navTextColor: #FFFFFF
-$navLinkWidth: 180px
-$hoverColor: #777
-
-// Active
-$activeColor: #F4F4F4
-$activeTextColor: #000000
-
-// Logout button
-$logoutBtnColor: #9AA8B5
-$logoutBtnTextColor: #000000
-$logoutBtnWidth: 180px
-$logoutBtnHeight: 32px
-
-// User settings button
-$transparentColor: rgba(0, 0, 0, 0)
-
